pyproject.toml
[tool.poetry]
name = "hk-flight-data"
version = "0.1.0"
description = ""
authors = ["adisbladis <adisbladis@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
tabula-py = "^2.3.0"
requests = "^2.27.1"
prettytable = "^3.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
